requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.1
pypdf>=3.15.1
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
flask>=2.2.3
gunicorn>=21.2.0
pytest>=7.3.1
numpy>=1.24.3
scikit-learn>=1.2.2
python-multipart>=0.0.6
tiktoken>=0.5.1
PyMuPDF>=1.22.5